requirements.txt
# InvoiceFlow Production Dependencies

# Core Framework
Django>=5.2.9
djangorestframework==3.15.2
drf-spectacular==0.28.0

# Database
dj-database-url==2.3.0
psycopg2-binary

# Security & Auth
django-csp==4.0
django-ratelimit==4.1.0
cryptography==46.0.3
pyotp==2.9.0
argon2-cffi

# Caching
django-redis==5.4.0

# Environment & Config
django-environ==0.12.0

# Web Server
gunicorn==23.0.0
whitenoise==6.11.0

# Email
sendgrid==6.12.5
python-http-client==3.3.7

# PDF Generation
weasyprint==66.0
reportlab==4.2.5
tinycss2==1.5.0
tinyhtml5==2.0.0
cssselect2==0.8.0
pydyf==0.11.0
pyphen==0.17.2
fonttools==4.61.0

# Image Processing
pillow==12.0.0
qrcode==7.4.2

# HTTP & Utils
requests==2.32.4
python-dateutil==2.9.0

# Compression
brotli==1.2.0
zopfli==0.4.0

# Monitoring
sentry-sdk==1.45.1
psutil

# Testing (dev)
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.10.0
factory-boy==3.3.1

# Misc
asgiref==3.11.0
cffi==2.0.0
pycparser==2.23
pytz==2024.1
PyYAML==6.0.3
sqlparse==0.5.3
webencodings==0.5.1
argon2-cffi
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==2.3.0
Django>=5.2.9
django-csp==4.0
django-environ==0.12.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
factory-boy==3.3.1
fonttools==4.61.0
gunicorn==23.0.0
pillow==12.0.0
psutil
psycopg2-binary
pycparser==2.23
pydyf==0.11.0
pyotp==2.9.0
pyphen==0.17.2
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.10.0
python-dateutil==2.9.0
python-http-client==3.3.7
pytz==2024.1
PyYAML==6.0.3
qrcode==7.4.2
reportlab==4.2.5
requests==2.32.4
sendgrid==6.12.5
sentry-sdk==1.45.1
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
argon2-cffi
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==2.3.0
Django>=5.2.9
django-csp==4.0
django-environ==0.12.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
factory-boy==3.3.1
fonttools==4.61.0
gunicorn==23.0.0
pillow==12.0.0
psutil
psycopg2-binary
pycparser==2.23
pydyf==0.11.0
pyotp==2.9.0
pyphen==0.17.2
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.10.0
python-dateutil==2.9.0
python-http-client==3.3.7
pytz==2024.1
PyYAML==6.0.3
qrcode==7.4.2
reportlab==4.2.5
requests==2.32.4
sendgrid==6.12.5
sentry-sdk==1.45.1
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
